+++ Editor/AssemblyInfo.cs
@@ -3,4 +3,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.PerformanceTesting.Tests.Editor")]
 
-[assembly: AssemblyVersion("3.2.0")]
+[assembly: AssemblyVersion("3.2.1")]